add: print_mutex

+++ src/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ecasalin <ecasalin@42.fr>                  +#+  +:+       +#+        */
+/*   By: ecasalin <ecasalin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/24 08:35:35 by ecasalin          #+#    #+#             */
-/*   Updated: 2025/06/24 11:04:10 by ecasalin         ###   ########.fr       */
+/*   Updated: 2025/06/25 10:06:39 by ecasalin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 long long	get_ms_diff(struct timeval start, struct timeval current)
 {
-	int	diff;
+	long long	diff;
 
 	diff = (current.tv_sec - start.tv_sec) * 1000;
 	diff += (current.tv_usec - start.tv_usec) / 1000;
